Move Enable Snap to Grid Button to Toolbox
Moved the enable snap to grid button from the details box to the toolbox.
I figured the best place to position the button was before the JSON file
download/upload buttons but after the tool buttons. This way the user has
quick access to snap to grid toggling when creating states and transitions.