Handle Y-Axis Overflow for Floating Panels
All floating panels are now scrollable when the content within each panel
extends past a max height. One thing to note, however, is that the Action
Stack floating panel will extend a little bit past the screen before becoming
scrollable, which is something that needs to be fixed. Also, it maybe worth
noting that the max height the Details Box obeys is a fixed number of pixels.